+++ src/main/java/uk/ac/ebi/protvar/model/data/CADDPrediction.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.protvar.model.response;
+package uk.ac.ebi.protvar.model.data;
 
 import lombok.Getter;
 import lombok.Setter;
