+++ src/main/java/uk/ac/ebi/protvar/model/data/ConservScore.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.protvar.model.response;
+package uk.ac.ebi.protvar.model.data;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
